Bump: Splits bump_clipfield tests by policy to improve build time/memory usage
